requirements.txt
absl-py==2.1.0
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
grpcio==1.62.1
importlib-metadata==7.0.2
importlib-resources==6.1.3
Jinja2==3.1.3
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1+cu118
typing-extensions==4.10.0
tzdata==2024.1
werkzeug==3.0.1
zipp==3.17.0
